requirements.txt
fastapi
pyotp
pydotenv
pyjwt